Move jsx to respective folders
