+++ docs/src/3.2.Camera.mdx
@@ -1,5 +1,5 @@
 import { CameraStateControlled, CameraStateUncontrolled } from './jsx/allLiveEditors'
-import CameraState from './jsx/CameraState.js';
+import CameraState from './jsx/api/CameraState.js';
 
 # Camera
