+++ docs/src/3.3.Command.mdx
@@ -1,4 +1,4 @@
-import CommandDemo from './jsx/Command.js';
+import CommandDemo from './jsx/api/Command.js';
 
 # Command
