--- docs/src/jsx/C22FollowObject.js
@@ -1,92 +1,0 @@
-//  Copyright (c) 118-present, GM Cruise LLC
-//
-//  This source code is licensed under the Apache License, Version 2.0,
-//  found in the LICENSE file in the root directory of this source tree.
-//  You may not use this file except in compliance with the License.
-
-// #BEGIN EXAMPLE
-import React from "react";
-
-import duckModel from "./utils/Duck.glb";
-import useRequestAnimationFrame from "./utils/useRequestAnimationFrame";
-import Worldview, { Spheres, Axes, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-worldview";
-
-// #BEGIN EDITABLE
-function Example() {
-  const steps = 500; // total amount of objects
-  const [count, setCount] = React.useState(0);
-  useRequestAnimationFrame(
-    () => {
-      // update count before each browser repaint
-      const newCount = (count + 1) % steps;
-      setCount(newCount);
-    },
-    false,
-    []
-  );
-
-  // map a number/index to a specific color
-  function numberToColor(number, max, a = 1) {
-    const i = (number * 255) / max;
-    const r = Math.round(Math.sin(0.024 * i + 0) * 127 + 128) / 255;
-    const g = Math.round(Math.sin(0.024 * i + 2) * 127 + 128) / 255;
-    const b = Math.round(Math.sin(0.024 * i + 4) * 127 + 128) / 255;
-    return { r, g, b, a };
-  }
-
-  // the object index needs to multiple by this scale so it's evenly distributed in the space
-  const scale = (Math.PI * 2) / steps;
-  const sphereMarker = {
-    pose: {
-      orientation: { x: 0, y: 0, z: 0, w: 1 },
-      position: { x: 0, y: 0, z: 0 },
-    },
-    scale: { x: 1, y: 1, z: 1 },
-    colors: [],
-    points: [],
-  };
-
-  new Array(steps)
-    .fill()
-    .map((_, idx) => [
-      // generate x, y, z coordinates based on trefoil equation
-      Math.sin(idx * scale) + 2 * Math.sin(2 * idx * scale),
-      Math.cos(idx * scale) - 2 * Math.cos(2 * idx * scale),
-      -Math.sin(3 * idx * scale),
-    ])
-    .forEach(([x, y, z], idx) => {
-      // add individual point and color to the single sphere object
-      sphereMarker.colors.push(numberToColor(idx, steps));
-      sphereMarker.points.push({ x: x * 20, y: y * 20, z: z * 20 });
-    });
-  const duckPosition = sphereMarker.points[count];
-
-  return (
-    <Worldview
-      cameraState={{
-        // Default setting for cameraState.
-        // Learn more at https://cruise-automation.github.io/webviz/worldview/#/docs/api/camera
-        ...DEFAULT_CAMERA_STATE,
-        // This is the magic! Simply supply the target position and the camera will follow
-        target: [duckPosition.x, duckPosition.y, duckPosition.z],
-        thetaOffset: -Math.PI / 2, // rotate the camera so the duck is facing right
-        // zoom out a little so we can see better
-        distance: 160,
-      }}>
-      <Spheres>{[sphereMarker]}</Spheres>
-      <Axes />
-      {/* Download model: https://github.com/cruise-automation/webviz/blob/master/docs/src/jsx/utils/Duck.glb  */}
-      <GLTFScene model={duckModel}>
-        {{
-          pose: {
-            position: duckPosition,
-            orientation: { x: 0, y: 0, z: 0, w: 1 },
-          },
-          scale: { x: 3, y: 3, z: 3 },
-        }}
-      </GLTFScene>
-    </Worldview>
-  );
-}
-// #END EXAMPLE
-export default Example;
