+++ docs/src/jsx/allDemos.stories.js
@@ -8,28 +8,28 @@ import { storiesOf } from "@storybook/react";
 import React from "react";
 import { withScreenshot } from "storybook-chrome-screenshot";
 
-import Arrows from "./Arrows";
-import BasicExample from "./BasicExample";
-import CameraStateControlled from "./CameraStateControlled";
-import CameraStateUncontrolled from "./CameraStateUncontrolled";
-import Composition from "./Composition";
-import Cones from "./Cones";
-import Cubes from "./Cubes";
-import Cylinders from "./Cylinders";
-import DuckScene from "./DuckScene";
-import DynamicCommands from "./DynamicCommands";
-import FilledPolygons from "./FilledPolygons";
-import Hitmap from "./Hitmap";
-import LinesDemo from "./LinesDemo";
-import LinesWireframe from "./LinesWireframe";
-import MouseEvents from "./MouseEvents";
-import Overlay from "./Overlay";
-import Points from "./Points";
-import SpheresInstanced from "./SpheresInstanced";
-import SpheresInstancedColor from "./SpheresInstancedColor";
-import SpheresSingle from "./SpheresSingle";
-import Text from "./Text";
-import Triangles from "./Triangles";
+import CameraStateControlled from "./api/CameraStateControlled";
+import CameraStateUncontrolled from "./api/CameraStateUncontrolled";
+import MouseEvents from "./api/MouseEvents";
+import Arrows from "./commands/Arrows";
+import Cones from "./commands/Cones";
+import Cubes from "./commands/Cubes";
+import Cylinders from "./commands/Cylinders";
+import DuckScene from "./commands/DuckScene";
+import FilledPolygons from "./commands/FilledPolygons";
+import LinesDemo from "./commands/LinesDemo";
+import LinesWireframe from "./commands/LinesWireframe";
+import Overlay from "./commands/Overlay";
+import Points from "./commands/Points";
+import SpheresInstanced from "./commands/SpheresInstanced";
+import SpheresInstancedColor from "./commands/SpheresInstancedColor";
+import SpheresSingle from "./commands/SpheresSingle";
+import Text from "./commands/Text";
+import Triangles from "./commands/Triangles";
+import BasicExample from "./examples/BasicExample";
+import Composition from "./examples/Composition";
+import DynamicCommands from "./examples/DynamicCommands";
+import Hitmap from "./examples/Hitmap";
 
 const allDemos = {
   Arrows,
