+++ docs/src/jsx/commands/Arrows.js
