+++ docs/src/jsx/commands/Cones.js
