+++ docs/src/jsx/commands/Cubes.js
