+++ docs/src/jsx/commands/DuckScene.js
