+++ docs/src/jsx/commands/FilledPolygons.js
@@ -8,7 +8,7 @@
 import polygonGenerator from "polygon-generator";
 import React from "react";
 
-import useRange from "./utils/useRange";
+import useRange from "../utils/useRange";
 import Worldview, { FilledPolygons, Axes } from "regl-worldview";
 
 // #BEGIN EDITABLE
