+++ docs/src/jsx/commands/FilledPolygonsHitmap.js
