+++ docs/src/jsx/commands/Overlay.js
@@ -7,7 +7,7 @@
 // #BEGIN EXAMPLE
 import React, { useState } from "react";
 
-import useRange from "./utils/useRange";
+import useRange from "../utils/useRange";
 import Worldview, { Axes, Overlay, Spheres } from "regl-worldview";
 
 // #BEGIN EDITABLE
