+++ docs/src/jsx/commands/Points.js
@@ -7,40 +7,44 @@
 // #BEGIN EXAMPLE
 import React from "react";
 
-import useRange from "./utils/useRange";
-import Worldview, { Spheres } from "regl-worldview";
+import useRange from "../utils/useRange";
+import Worldview, { Axes, Points, DEFAULT_CAMERA_STATE } from "regl-worldview";
 
 // #BEGIN EDITABLE
 function Example() {
   const range = useRange();
   // create coordinates
-  const x = 20;
-  const y = 20;
-  const z = 20;
+  const x = 3 + range * 2;
+  const y = x;
+  const z = x;
   const step = 10;
-  const coords = [];
+  const points = [];
   for (let i = 0; i < x; i++) {
     for (let j = 0; j < y; j++) {
       for (let k = 0; k < z; k++) {
-        coords.push({ x: i * step, y: j * step, z: k * step });
+        points.push({ x: i * step, y: j * step, z: k * step });
       }
     }
   }
-  // create sphere list
-  const scaleX = 0.25 * (1 + range);
-  const sphereList = {
-    points: coords,
+
+  const scaleX = 3 * range;
+  const marker = {
+    points,
     scale: { x: scaleX, y: scaleX, z: scaleX },
     color: { r: 1, g: range, b: 1, a: 1 },
-    pose: {
-      position: { x: range, y: range, z: range },
-      orientation: { x: 0, y: 0, z: 0, w: 1 },
-    },
+    pose: { position: { x: range, y: range, z: range }, orientation: { x: 0, y: 0, z: 0, w: 1 } },
   };
 
   return (
-    <Worldview>
-      <Spheres>{[sphereList]}</Spheres>
+    <Worldview
+      defaultCameraState={{
+        ...DEFAULT_CAMERA_STATE,
+        distance: 124,
+        phi: 1,
+        targetOffset: [3, 6, 0],
+      }}>
+      <Points>{[marker]}</Points>
+      <Axes />
     </Worldview>
   );
 }
