+++ docs/src/jsx/commands/SpheresInstancedColor.js
