+++ docs/src/jsx/commands/SpheresSingle.js
