+++ docs/src/jsx/commands/Text.js
