+++ docs/src/jsx/commands/Triangles.js
