+++ docs/src/jsx/examples/BasicExample.js
