+++ docs/src/jsx/examples/Composition.js
