+++ docs/src/jsx/examples/Hitmap.js
