+++ docs/src/jsx/tutorials/C12CustomObject.js
