+++ docs/src/jsx/tutorials/C14InstancedRendering.js
