+++ docs/src/jsx/tutorials/C23FollowObjectOrientation.js
@@ -5,26 +5,24 @@
 //  You may not use this file except in compliance with the License.
 
 // #BEGIN EXAMPLE
+import { quat, vec3 } from "gl-matrix";
 import React from "react";
 
-import duckModel from "./utils/Duck.glb";
-import useRequestAnimationFrame from "./utils/useRequestAnimationFrame";
-import Worldview, { Cubes, Spheres, Axes, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-worldview";
+import duckModel from "../utils/Duck.glb";
+import useRequestAnimationFrame from "../utils/useRequestAnimationFrame";
+import Worldview, { Spheres, Axes, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-worldview";
 
 // #BEGIN EDITABLE
 function Example() {
   const steps = 500; // total amount of objects
-  const [clickedObjectIds, setClickedObjectIds] = React.useState([]);
   const [count, setCount] = React.useState(0);
-  const [shouldStopDuck, setShouldStopDuck] = React.useState(false);
-
   useRequestAnimationFrame(
     () => {
       // update count before each browser repaint
       const newCount = (count + 1) % steps;
       setCount(newCount);
     },
-    shouldStopDuck, // disable the animation when the duck is not moving
+    false,
     []
   );
 
@@ -40,12 +38,11 @@ function Example() {
   // the object index needs to multiple by this scale so it's evenly distributed in the space
   const scale = (Math.PI * 2) / steps;
   const sphereMarker = {
-    id: 1001,
     pose: {
       orientation: { x: 0, y: 0, z: 0, w: 1 },
       position: { x: 0, y: 0, z: 0 },
     },
-    scale: { x: 3, y: 3, z: 3 },
+    scale: { x: 1, y: 1, z: 1 },
     colors: [],
     points: [],
   };
@@ -63,68 +60,48 @@ function Example() {
       sphereMarker.colors.push(numberToColor(idx, steps));
       sphereMarker.points.push({ x: x * 20, y: y * 20, z: z * 20 });
     });
-
-  const obstacleMarkers = Array.from(clickedObjectIds).map((clickedObjectId, index) => {
-    const pointIdx = clickedObjectId - sphereMarker.id;
-    const position = sphereMarker.points[pointIdx];
-    return {
-      // Since the `sphereMarker` has used up the id range: 101 ~ 101 + 499 (inclusive, each id represent one sphere object),
-      // to make the obstacleMarkers' ids unique, we'll use the range: 500 (sphereMarker.id + step) + index.
-      // Learn about id mapping at https://cruise-automation.github.io/webviz/worldview/#/docs/api/mouse-events
-      id: sphereMarker.id + steps + index,
-      // remember the original clickedObjectId so when the obstacle is clicked, we can
-      // remove the obstacle quickly by updating clickedObjectIds
-      clickedObjectId,
-      pose: {
-        orientation: { x: 0, y: 0, z: 0, w: 1 },
-        position,
-      },
-      color: { r: 1, g: 0, b: 0, a: 1 }, // red
-      scale: { x: 6, y: 6, z: 6 }, // scale up a little so it's bigger than the spheres
-    };
-  });
   const duckPosition = sphereMarker.points[count];
-  // make the duck stop
-  React.useEffect(
-    () => {
-      const duckPositionId = sphereMarker.id + count;
-      if (!shouldStopDuck && clickedObjectIds.includes(duckPositionId)) {
-        setShouldStopDuck(true);
-      }
-    },
-    [clickedObjectIds, shouldStopDuck, count, sphereMarker.id]
+
+  // get the orientation for the duck so its moving direction is always aligned with the knot
+  const tangentVecO = [
+    Math.cos(count * scale) + 4 * Math.cos(2 * count * scale),
+    -Math.sin(count * scale) + 4 * Math.sin(2 * count * scale),
+    -3 * Math.cos(3 * count * scale),
+  ];
+  const tangentVec = vec3.normalize([0, 0, 0], tangentVecO);
+  const duckYaw = quat.setAxisAngle([0, 0, 0, 0], [0, 0, 1], Math.atan2(tangentVec[1], tangentVec[0]));
+  const duckOrientation = [0, 0, 0, 1];
+  quat.multiply(duckOrientation, duckOrientation, duckYaw);
+  quat.rotateY(
+    duckOrientation,
+    duckOrientation,
+    -Math.atan(tangentVecO[2] / Math.hypot(tangentVecO[0], tangentVecO[1]))
   );
+  quat.rotateZ(duckOrientation, duckOrientation, -Math.PI / 2);
 
   return (
     <Worldview
-      defaultCameraState={{
+      cameraState={{
+        // Default setting for cameraState.
+        // Learn more at https://cruise-automation.github.io/webviz/worldview/#/docs/api/camera
         ...DEFAULT_CAMERA_STATE,
+        target: [duckPosition.x, duckPosition.y, duckPosition.z],
+        // This is the magic! The `targetOrientation` input will make sure the camera follows the duck's orientation
+        targetOrientation: duckOrientation,
+        // zoom out a little so we can see better
         distance: 160,
         thetaOffset: -Math.PI / 2, // rotate the camera so the duck is facing right
       }}>
-      <Spheres
-        onClick={(ev, { objectId }) => {
-          setClickedObjectIds([...clickedObjectIds, objectId]);
-        }}>
-        {[sphereMarker]}
-      </Spheres>
-      <Cubes
-        onClick={(ev, { object }) => {
-          const newClickedObjectIds = clickedObjectIds.filter((id) => id !== object.clickedObjectId);
-          setClickedObjectIds(newClickedObjectIds);
-          setShouldStopDuck(false);
-        }}>
-        {obstacleMarkers}
-      </Cubes>
+      <Spheres>{[sphereMarker]}</Spheres>
       <Axes />
       {/* Download model: https://github.com/cruise-automation/webviz/blob/master/docs/src/jsx/utils/Duck.glb  */}
       <GLTFScene model={duckModel}>
         {{
           pose: {
             position: duckPosition,
-            orientation: { x: 0, y: 0, z: 0, w: 1 },
+            orientation: { x: duckOrientation[0], y: duckOrientation[1], z: duckOrientation[2], w: duckOrientation[3] },
           },
-          scale: { x: 4, y: 4, z: 4 },
+          scale: { x: 3, y: 3, z: 3 },
         }}
       </GLTFScene>
     </Worldview>
